+++ src/main/java/ie/atu/userinterface/Admin/Admin.java
@@ -1,0 +1,13 @@
+package ie.atu.userinterface.Admin;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Admin {
+    private String username;
+    private String password;
+}
