+++ src/main/java/ie/atu/userinterface/GPU/GPUService.java
@@ -1,12 +1,6 @@
 package ie.atu.userinterface.GPU;
 
-import ie.atu.userinterface.CPU.CPU;
-import ie.atu.userinterface.CompatibilityRequest;
-import ie.atu.userinterface.RAM.RAM;
-import ie.atu.userinterface.RAM.RAMClient;
-import ie.atu.userinterface.RAM.RAMHardwareClient;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
