+++ src/main/java/ie/atu/userinterface/Motherboard/MotherboardController.java
@@ -1,7 +1,7 @@
 package ie.atu.userinterface.Motherboard;
 
 import ie.atu.userinterface.CPU.CPU;
-import ie.atu.userinterface.CompatibilityRequest;
+import ie.atu.userinterface.Compatibility.CompatibilityRequest;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
