+++ src/main/java/ie/atu/userinterface/Web/LoginClient.java
@@ -1,5 +1,6 @@
 package ie.atu.userinterface.Web;
 
+import ie.atu.userinterface.Admin.Admin;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
