+++ src/main/java/ie/atu/userinterface/Web/WebService.java
@@ -1,5 +1,6 @@
 package ie.atu.userinterface.Web;
 
+import ie.atu.userinterface.Admin.Admin;
 import org.springframework.stereotype.Service;
 
 @Service
